srv/blackroad/requirements.txt
Flask>=3.0
requests>=2.31
python-dotenv>=1.0
flask-cors>=4.0
boto3>=1.34
